♻️ Refactor api codes by given url
Refactor aoi codes due to api changes
