+++ pages/api/NavBarApi.ts
@@ -1,9 +1,9 @@
 import { DEFAULT_PROFILE } from "@/src/util/constant";
-import instance from "@/src/util/instance";
+import instance from "@/pages/api/instance";
 
 export const getUser = async () => {
   const response = await instance.get("users");
-  const data = response.data ? response.data?.data[0] : null;
+  const data = response?.data[0];
   const userData = data
     ? {
         id: data.id,
@@ -12,5 +12,6 @@ export const getUser = async () => {
         profileImageSource: data.image_source || DEFAULT_PROFILE,
       }
     : null;
+
   return userData;
 };
